+++ CHANGELOG.md
@@ -5,3 +5,6 @@
 
 ## 0.2.0 - 2016-09-04
 - Allow selecting specific fields from hash.get and hash.search
+
+## 0.3.0 - 2016-09-04
+- Allow removing specified hash by key